+++ main-server/src/core/common/index.js
@@ -23,7 +23,6 @@
 import EventEmitter from 'events';
 
 import CustomError from 'ndid-error/custom_error';
-import errorType from 'ndid-error/type';
 import logger from '../../logger';
 
 import * as tendermint from '../../tendermint';
@@ -35,13 +34,11 @@ import * as lt from '../../utils/long_timeout';
 import * as config from '../../config';
 import { getErrorObjectForClient } from '../../utils/error';
 import * as cacheDb from '../../db/cache';
-import privateMessageType from '../../mq/message/type';
 
 import { delegateToWorker } from '../../master-worker-interface/server';
 import { broadcastRemoveRequestTimeoutScheduler } from '../../master-worker-interface/client';
 
 import MODE from '../../mode';
-import ROLE from '../../role';
 import { getFunction } from '../../functions';
 
 export * from './create_request';
@@ -170,6 +167,7 @@ export async function getIdpMQDestinations({
   min_aal,
   idp_id_list,
   mode,
+  supported_request_message_data_url_type_list,
 }) {
   const idpNodes = await tendermintNdid.getIdpNodesInfo({
     namespace: mode === 2 || mode === 3 ? namespace : undefined,
@@ -177,6 +175,8 @@ export async function getIdpMQDestinations({
     min_ial,
     min_aal,
     node_id_list: idp_id_list, // filter to include only nodes in this list if node ID exists
+    supported_request_message_data_url_type_list, // filter to include only nodes that support all request message types in this list
+    mode_list: mode === 2 || mode === 3 ? [mode] : undefined, //  filter to include only nodes that support all modes in this list
   });
 
   const receivers = idpNodes