+++ main-server/src/http_server/routes/validator/json_schema/ndid.js
@@ -167,8 +167,24 @@ export default {
         properties: {
           namespace: { type: 'string', minLength: 1 },
           description: { type: 'string' },
+          allowed_identifier_count_in_reference_group: {
+            type: 'integer',
+            minimum: -1,
+          },
         },
         required: ['namespace'],
+      },
+    },
+    '/ndid/namespaces/:namespace': {
+      body: {
+        $schema: 'http://json-schema.org/draft-07/schema#',
+        properties: {
+          description: { type: 'string' },
+          allowed_identifier_count_in_reference_group: {
+            type: 'integer',
+            minimum: -1,
+          },
+        },
       },
     },
     '/ndid/services': {
@@ -274,7 +290,7 @@ export default {
       body: {
         $schema: 'http://json-schema.org/draft-07/schema#',
         properties: {
-          purpose: { type: 'string', minLength: 1 },
+          purpose: { type: 'string' },
           allowed_mode_list: {
             type: 'array',
             items: {